Support the full scanner image size
When querying for the scanner capabilities capture the full height and
width for both the platen and ADF. Then pass these as the height and
width for the job settings and update the job XML.

This allows for scans to scan up to the full size available for the
input source rather than the hard-coded size that was previously used.
This PR also removes the 33 pixel cutoff from the X axis.